// src/lib/openapi/squad/models/CreateOutcomePayload.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Squad API
 * API for managing Squad resources
 *
 * The version of the OpenAPI document: 0.14.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime.js';
/**
 * Request schema for creating an outcome
 * @export
 * @interface CreateOutcomePayload
 */
export interface CreateOutcomePayload {
    /**
     * 
     * @type {string}
     * @memberof CreateOutcomePayload
     */
    id?: string;
    /**
     * Title of the outcome
     * @type {string}
     * @memberof CreateOutcomePayload
     */
    title: string;
    /**
     * Detailed description of the outcome
     * @type {string}
     * @memberof CreateOutcomePayload
     */
    description: string;
    /**
     * Priority level of the outcome
     * @type {number}
     * @memberof CreateOutcomePayload
     */
    priority: number;
    /**
     * Trend indicator for the outcome
     * @type {number}
     * @memberof CreateOutcomePayload
     */
    trend?: number;
    /**
     * List of analytic events associated with the outcome
     * @type {Array<string>}
     * @memberof CreateOutcomePayload
     */
    analyticEvents?: Array<string>;
    /**
     * ID of the owner of the outcome
     * @type {string}
     * @memberof CreateOutcomePayload
     */
    ownerId?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateOutcomePayload
     */
    opportunityIds?: Array<string>;
}

/**
 * Check if a given object implements the CreateOutcomePayload interface.
 */
export function instanceOfCreateOutcomePayload(value: object): value is CreateOutcomePayload {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('priority' in value) || value['priority'] === undefined) return false;
    return true;
}

export function CreateOutcomePayloadFromJSON(json: any): CreateOutcomePayload {
    return CreateOutcomePayloadFromJSONTyped(json, false);
}

export function CreateOutcomePayloadFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateOutcomePayload {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'],
        'description': json['description'],
        'priority': json['priority'],
        'trend': json['trend'] == null ? undefined : json['trend'],
        'analyticEvents': json['analyticEvents'] == null ? undefined : json['analyticEvents'],
        'ownerId': json['ownerId'] == null ? undefined : json['ownerId'],
        'opportunityIds': json['opportunityIds'] == null ? undefined : json['opportunityIds'],
    };
}

export function CreateOutcomePayloadToJSON(json: any): CreateOutcomePayload {
    return CreateOutcomePayloadToJSONTyped(json, false);
}

export function CreateOutcomePayloadToJSONTyped(value?: CreateOutcomePayload | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'description': value['description'],
        'priority': value['priority'],
        'trend': value['trend'],
        'analyticEvents': value['analyticEvents'],
        'ownerId': value['ownerId'],
        'opportunityIds': value['opportunityIds'],
    };
}